@@ -1,7 +1,7 @@
 [package]
 name = "mcp-bridge"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [lib]
 name = "mcp_bridge"
